// types.ts
export enum AppView {
    EXPLORE = 'EXPLORE',
    TRIPS = 'TRIPS',
    INBOX = 'INBOX',
    PROFILE = 'PROFILE',
    SAFETY = 'SAFETY',
    AI_PLANNER = 'AI_PLANNER',
    BADGES = 'BADGES'
}
// ...
export interface UserPrefs {
    travelMode: 'Private' | 'Metro' | 'Walk';
    dietary: string[];
}

export interface Activity {
    id: string;
    title: string;
    location: string;
    price: number;
    rating: number;
    reviews: string;
    image: string;
    category: string;
    duration: string;
    tags: string[];
}